split code into files and create analyzation
