+++ scaling-test/profiler.py
@@ -3,16 +3,19 @@ Adapted from: https://github.com/pytorch/examples/blob/main/mnist/main.py
 """
 
 import argparse
-import types
+from pathlib import Path
+import torch
+import pandas as pd
 
-import torchvision
-from dataset import imagenet_dataset
-from torch.utils.data import Subset
+from torchvision import datasets, transforms
 
-from itwinai.components import profile_torch_trainer
 from itwinai.loggers import ConsoleLogger
 from itwinai.torch.config import TrainingConfiguration
 from itwinai.torch.trainer import TorchTrainer
+
+from net import Net
+from torch.profiler import ProfilerActivity, profile
+
 
 
 def main():
@@ -21,13 +24,8 @@ def main():
     parser.add_argument(
         "--batch-size",
         type=int,
-        default=128,
+        default=64,
         help="input batch size for training (default: 64)",
-    )
-    parser.add_argument(
-        "--data-dir",
-        default="/p/scratch/intertwin/datasets/imagenet/ILSVRC2012/train/",
-        help=("location of the training dataset in the local filesystem"),
     )
     parser.add_argument(
         "--epochs", type=int, default=2, help="number of epochs to train (default: 14)"
@@ -35,75 +33,77 @@ def main():
     parser.add_argument(
         "--strategy", type=str, default="ddp", help="distributed strategy (default=ddp)"
     )
+    parser.add_argument(
+        "--lr", type=float, default=1e-4, help="learning rate (default: 1.0)"
+    )
+    parser.add_argument("--seed", type=int, default=1, help="random seed (default: 1)")
+    parser.add_argument(
+        "--ckpt-interval",
+        type=int,
+        default=10,
+        help="how many batches to wait before logging training status",
+    )
     args = parser.parse_args()
-    train_dataset = imagenet_dataset(args.data_dir)
 
-    # Only consider a subset of the data
-    indices = list(range(1500))
-    train_dataset = Subset(train_dataset, indices)
+    # Dataset creation
+    transform = transforms.Compose(
+        [transforms.ToTensor(), transforms.Normalize((0.1307,), (0.3081,))]
+    )
+    root_dir = Path("data")
+    train_dataset = datasets.MNIST(
+        str(root_dir), train=True, download=True, transform=transform
+    )
+    validation_dataset = datasets.MNIST(str(root_dir), train=False, transform=transform)
 
-    model = torchvision.models.resnet152()
+    # Neural network to train
+    model = Net()
 
     training_config = TrainingConfiguration(
         batch_size=args.batch_size,
-        optimizer="sgd",
-        loss="nllloss",
-        num_workers_dataloader=1,
+        optim_lr=args.lr,
+        optimizer="adadelta",
+        loss="cross_entropy",
+        num_workers_dataloader=1
     )
     logger = ConsoleLogger()
 
-    # profiler = profile(
-    #     activities=[ProfilerActivity.CUDA, ProfilerActivity.CPU],
-    #     with_modules=True,
-    # )
+    profiler = profile(
+        activities=[ProfilerActivity.CUDA, ProfilerActivity.CPU], 
+        # on_trace_ready=torch.profiler.tensorboard_trace_handler("./logs"),
+        experimental_config=torch._C._profiler._ExperimentalConfig(verbose=True),
+        record_shapes=True,
+        # profile_memory=True,
+    )
 
     trainer = TorchTrainer(
         config=training_config,
         model=model,
         strategy=args.strategy,
         epochs=args.epochs,
-        logger=logger,
-        validation_every=None,
+        random_seed=args.seed,
+        checkpoint_every=args.ckpt_interval,
+        logger=logger
     )
 
-    # strategy = trainer.strategy
-    trainer.execute = types.MethodType(profile_torch_trainer(trainer.execute), trainer)
+    profiler.start()
+    trainer.execute(train_dataset, validation_dataset, profiler=profiler)
+    profiler.stop()
 
-    # profiler.start()
-    trainer.execute(train_dataset)
-    # profiler.stop()
+    print(f"Running this!")
 
-    # key_averages = profiler.key_averages(group_by_stack_n=0)
-    # key_averages = profiler.key_averages()
-    # print(key_averages.table(sort_by="cuda_time_total"))
-    #
-    # profiling_data = []
-    # for event in key_averages:
-    #     profiling_data.append(
-    #         {
-    #             "name": event.key,
-    #             "node_id": event.node_id,
-    #             "self_cpu_time_total": event.self_cpu_time_total,
-    #             "cpu_time_total": event.cpu_time_total,
-    #             "cpu_time_total_str": event.cpu_time_total_str,
-    #             "self_cuda_time_total": event.self_cuda_time_total,
-    #             "cuda_time_total": event.cuda_time_total,
-    #             "cuda_time_total_str": event.cuda_time_total_str,
-    #             "calls": event.count,
-    #         }
-    #     )
-    # grank = strategy.global_rank()
-    # num_gpus = strategy.global_world_size()
-    # df = pd.DataFrame(profiling_data)
-    #
-    # output_dir = Path("profiling_logs")
-    # output_dir.mkdir(parents=True, exist_ok=True)
-    #
-    # output_path = output_dir / f"profile_{args.strategy}_{num_gpus}_{grank}.csv"
-    #
-    # print(f"Saving the dataframe to {output_path}", force=True)
-    # df.to_csv(output_path, index=False)
-    # strategy.clean_up()
+    profiling_data = []
+    for event in profiler.key_averages(): 
+        profiling_data.append({
+        "name": event.key,
+        "cpu_time": event.cpu_time_total,
+        "self_cpu_time": event.self_cpu_time_total,
+        "gpu_time": event.cuda_time_total,
+        "self_gpu_time": event.self_cuda_time_total,
+        "calls": event.count,
+    })
+
+    df = pd.DataFrame(profiling_data)
+    df.to_csv("logs/dataframe.csv")
 
 
 if __name__ == "__main__":
